+++ sentry-kotlin-multiplatform/src/commonStub/kotlin/io/sentry/kotlin/multiplatform/SentryBridge.commonStub.kt
@@ -1,6 +1,7 @@
 package io.sentry.kotlin.multiplatform
 
 import io.sentry.kotlin.multiplatform.log.SentryLogger
+import io.sentry.kotlin.multiplatform.metrics.SentryMetrics
 import io.sentry.kotlin.multiplatform.protocol.Breadcrumb
 import io.sentry.kotlin.multiplatform.protocol.SentryId
 import io.sentry.kotlin.multiplatform.protocol.User
@@ -54,6 +55,10 @@ internal actual class SentryBridge actual constructor(
     actual fun setUser(user: User?) {
         // No-op
     }
+
+    private val metrics: SentryMetrics = NoOpSentryMetrics
+
+    actual fun metrics(): SentryMetrics = metrics
 
     actual fun logger(): SentryLogger = NoOpSentryLogger()